feat: Added ability to send hearts without using script
